Refactor RepListCommand, delete old coinflip bet
