+++ loritta-plugins/helping-hands/src/main/kotlin/net/perfectdreams/loritta/plugin/helpinghands/HelpingHandsPlugin.kt
@@ -4,18 +4,20 @@ import mu.KotlinLogging
 import net.perfectdreams.loritta.api.LorittaBot
 import net.perfectdreams.loritta.platform.discord.LorittaDiscord
 import net.perfectdreams.loritta.platform.discord.plugin.LorittaDiscordPlugin
-import net.perfectdreams.loritta.plugin.helpinghands.commands.*
+import net.perfectdreams.loritta.plugin.helpinghands.commands.CoinFlipBetCommand
+import net.perfectdreams.loritta.plugin.helpinghands.commands.DailyInactivityTaxExecutor
+import net.perfectdreams.loritta.plugin.helpinghands.commands.RepListCommand
+import net.perfectdreams.loritta.plugin.helpinghands.commands.TestCommand
 import net.perfectdreams.loritta.plugin.helpinghands.utils.DailyInactivityTaxUtils
 
 class HelpingHandsPlugin(name: String, loritta: LorittaBot) : LorittaDiscordPlugin(name, loritta) {
 	override fun onEnable() {
 		loritta as LorittaDiscord
-
-		registerCommand(CoinFlipBetCommand.command(this, loritta))
-		registerCommand(RepListCommand.create(loritta))
+		
 		registerCommands(
-				TestCommand(this),
-				CoinFlipBetXCommand(this)
+				CoinFlipBetCommand(this),
+				RepListCommand(this),
+				TestCommand(this)
 		)
 
 		if (loritta.isMaster)
