+++ loritta-plugins/helping-hands/src/main/kotlin/net/perfectdreams/loritta/plugin/helpinghands/commands/CoinFlipBetCommand.kt
@@ -3,32 +3,35 @@ package net.perfectdreams.loritta.plugin.helpinghands.commands
 import com.mrpowergamerbr.loritta.Loritta
 import com.mrpowergamerbr.loritta.commands.vanilla.`fun`.CaraCoroaCommand
 import com.mrpowergamerbr.loritta.utils.Constants
-import com.mrpowergamerbr.loritta.utils.MessageInteractionFunctions
+import com.mrpowergamerbr.loritta.utils.onReactionAdd
 import com.mrpowergamerbr.loritta.utils.removeAllFunctions
 import com.mrpowergamerbr.loritta.utils.stripCodeMarks
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import net.dv8tion.jda.api.entities.User
-import net.perfectdreams.loritta.api.LorittaBot
 import net.perfectdreams.loritta.api.commands.ArgumentType
+import net.perfectdreams.loritta.api.commands.CommandCategory
 import net.perfectdreams.loritta.api.commands.arguments
 import net.perfectdreams.loritta.api.messages.LorittaReply
-import net.perfectdreams.loritta.platform.discord.commands.DiscordCommandContext
+import net.perfectdreams.loritta.platform.discord.commands.DiscordAbstractCommandBase
 import net.perfectdreams.loritta.plugin.helpinghands.HelpingHandsPlugin
-import net.perfectdreams.loritta.plugin.helpinghands.commands.base.DSLCommandBase
 import net.perfectdreams.loritta.utils.*
 import net.perfectdreams.loritta.utils.extensions.refreshInDeferredTransaction
 import net.perfectdreams.loritta.utils.extensions.toJDA
 
-object CoinFlipBetCommand : DSLCommandBase {
-	val mutex = Mutex()
+class CoinFlipBetCommand(val plugin: HelpingHandsPlugin) : DiscordAbstractCommandBase(
+		plugin.lorittaDiscord,
+		listOf("coinflip", "flipcoin", "girarmoeda", "caracoroa")
+				.flatMap { listOf("$it bet", "$it apostar") },
+		CommandCategory.ECONOMY
+) {
+	companion object {
+		// Used to avoid dupes
+		private val mutex = Mutex()
+	}
 
-	override fun command(plugin: HelpingHandsPlugin, loritta: LorittaBot) = create(
-			loritta,
-			listOf("coinflip", "flipcoin", "girarmoeda", "caracoroa")
-					.flatMap { listOf("$it bet", "$it apostar") }
-	) {
+	override fun command() = create {
 		localizedDescription("commands.economy.flipcoinbet.description")
 
 		examples {
@@ -45,24 +48,18 @@ object CoinFlipBetCommand : DSLCommandBase {
 
 		this.canUseInPrivateChannel = false
 
-		executes {
-			loritta as Loritta
+		executesDiscord {
+			if (2 > args.size)
+				this.explainAndExit()
 
-			if (2 > args.size) {
-				this.explain()
-				return@executes
-			}
-
-			val context = checkType<DiscordCommandContext>(this)
-
-			val _user = validate(context.user(0))
+			val _user = validate(user(0))
 			val invitedUser = _user.toJDA()
 
-			if (invitedUser == context.user)
+			if (invitedUser == user)
 				fail(locale["commands.economy.flipcoinbet.cantBetSelf"], Constants.ERROR)
 
-			val selfActiveDonations = com.mrpowergamerbr.loritta.utils.loritta.getActiveMoneyFromDonationsAsync(context.discordMessage.author.idLong)
-			val otherActiveDonations = com.mrpowergamerbr.loritta.utils.loritta.getActiveMoneyFromDonationsAsync(invitedUser.idLong)
+			val selfActiveDonations = loritta.getActiveMoneyFromDonationsAsync(discordMessage.author.idLong)
+			val otherActiveDonations = loritta.getActiveMoneyFromDonationsAsync(invitedUser.idLong)
 
 			val selfPlan = UserPremiumPlans.getPlanFromValue(selfActiveDonations)
 			val otherPlan = UserPremiumPlans.getPlanFromValue(otherActiveDonations)
@@ -72,7 +69,7 @@ object CoinFlipBetCommand : DSLCommandBase {
 			val plan: UserPremiumPlans?
 
 			if (selfPlan.totalCoinFlipReward == 1.0) {
-				whoHasTheNoTaxReward = context.discordMessage.author
+				whoHasTheNoTaxReward = discordMessage.author
 				hasNoTax = true
 				plan = selfPlan
 			} else if (otherPlan.totalCoinFlipReward == 1.0) {
@@ -85,8 +82,8 @@ object CoinFlipBetCommand : DSLCommandBase {
 				plan = UserPremiumPlans.Essential
 			}
 
-			val number = NumberUtils.convertShortenedNumberToLong(context.args[1])
-					?: fail(locale["commands.invalidNumber", context.args[1].stripCodeMarks()], Emotes.LORI_CRYING.toString())
+			val number = NumberUtils.convertShortenedNumberToLong(args[1])
+					?: fail(locale["commands.invalidNumber", args[1].stripCodeMarks()], Emotes.LORI_CRYING.toString())
 			val tax = (number * (1.0 - plan.totalCoinFlipReward)).toLong()
 			val money = number - tax
 
@@ -96,7 +93,7 @@ object CoinFlipBetCommand : DSLCommandBase {
 			if (0 >= number)
 				fail(locale["commands.economy.flipcoinbet.zeroMoney"], Constants.ERROR)
 
-			val selfUserProfile = context.lorittaUser.profile
+			val selfUserProfile = lorittaUser.profile
 
 			if (number > selfUserProfile.money)
 				fail(locale["commands.economy.flipcoinbet.notEnoughMoneySelf"], Constants.ERROR)
@@ -107,38 +104,33 @@ object CoinFlipBetCommand : DSLCommandBase {
 			if (number > invitedUserProfile.money || bannedState != null)
 				fail(locale["commands.economy.flipcoinbet.notEnoughMoneyInvited", invitedUser.asMention], Constants.ERROR)
 
-			val message = context.reply(
-					LorittaReply(
-							(
-									if (hasNoTax)
-										locale[
-												"commands.economy.flipcoinbet.startBetNoTax",
-												invitedUser.asMention,
-												context.user.asMention,
-												locale["commands.fun.flipcoin.heads"],
-												money,
-												locale["commands.fun.flipcoin.tails"],
-												whoHasTheNoTaxReward?.asMention ?: "???"
-										]
-									else
-										locale[
-												"commands.economy.flipcoinbet.startBet",
-												invitedUser.asMention,
-												context.user.asMention,
-												locale["commands.fun.flipcoin.heads"],
-												money,
-												locale["commands.fun.flipcoin.tails"],
-												number,
-												tax
-										]
-									),
-							Emotes.LORI_RICH,
-							mentionUser = false
-					)
+			val message = reply(
+					if (hasNoTax)
+						locale[
+								"commands.economy.flipcoinbet.startBetNoTax",
+								invitedUser.asMention,
+								user.asMention,
+								locale["commands.fun.flipcoin.heads"],
+								money,
+								locale["commands.fun.flipcoin.tails"],
+								whoHasTheNoTaxReward?.asMention ?: "???"
+						]
+					else
+						locale[
+								"commands.economy.flipcoinbet.startBet",
+								invitedUser.asMention,
+								user.asMention,
+								locale["commands.fun.flipcoin.heads"],
+								money,
+								locale["commands.fun.flipcoin.tails"],
+								number,
+								tax
+						],
+					Emotes.LORI_RICH,
+					mentionUser = false
 			).toJDA()
 
-			val functions = com.mrpowergamerbr.loritta.utils.loritta.messageInteractionCache.getOrPut(message.idLong) { MessageInteractionFunctions(message.guild.idLong, message.channel.idLong, context.user.idLong) }
-			functions.onReactionAdd = {
+			message.onReactionAdd(this) {
 				if (it.userIdLong == invitedUser.idLong && it.reactionEmote.name == "✅") {
 					message.removeAllFunctions()
 					plugin.launch {
@@ -160,17 +152,17 @@ object CoinFlipBetCommand : DSLCommandBase {
 
 							if (isTails) {
 								prefix = "<:coroa:412586257114464259>"
-								message = context.locale["${CaraCoroaCommand.LOCALE_PREFIX}.tails"]
+								message = locale["${CaraCoroaCommand.LOCALE_PREFIX}.tails"]
 							} else {
 								prefix = "<:cara:412586256409559041>"
-								message = context.locale["${CaraCoroaCommand.LOCALE_PREFIX}.heads"]
+								message = locale["${CaraCoroaCommand.LOCALE_PREFIX}.heads"]
 							}
 
-							val winner: net.dv8tion.jda.api.entities.User
-							val loser: net.dv8tion.jda.api.entities.User
+							val winner: User
+							val loser: User
 
 							if (isTails) {
-								winner = context.user
+								winner = user
 								loser = invitedUser
 								loritta.newSuspendedTransaction {
 									selfUserProfile.addSonhosNested(money)
@@ -185,7 +177,7 @@ object CoinFlipBetCommand : DSLCommandBase {
 								}
 							} else {
 								winner = invitedUser
-								loser = context.user
+								loser = user
 								loritta.newSuspendedTransaction {
 									invitedUserProfile.addSonhosNested(money)
 									selfUserProfile.takeSonhosNested(number)
@@ -199,7 +191,7 @@ object CoinFlipBetCommand : DSLCommandBase {
 								}
 							}
 
-							context.reply(
+							reply(
 									LorittaReply(
 											"**$message!**",
 											prefix,
